More type cleanup to squelch warnings.
Make exported_pdu_headers_size a size_t to make it big enough.  Make
used_buffer_length a size_t as it doesn't go negative and gets compared
with unsigned values.  Cast length to size_t when comparing it with
used_buffer_length - by that time, we've ensured that it's not negative.

Change-Id: Ie192eee55914ebe548865d46575768964353ea2d
Reviewed-on: https://code.wireshark.org/review/8021
Reviewed-by: Guy Harris <guy@alum.mit.edu>