Simplify the calculation of the time stamp.
The time is calculated based on a 32-bit "seconds since the Epoch" value
for the start time and a 32-bit delta from that time, in milliseconds.
We can just split that delta into seconds and milliseconds, add the
seconds to the start time to get the seconds part of the time stamp, and
multiply the milliseconds by 1,000,000 to get the nanoseconds part of
the time stamp.  The only 64-bit arithmetic needed is adding the seconds
to a 64-bit version of the start time (just in case seconds+start time
goes past Y2.038K).

Change-Id: I6b3319ec73f1fd49dd59435d0a2ccc2365edf986
Reviewed-on: https://code.wireshark.org/review/8156
Reviewed-by: Guy Harris <guy@alum.mit.edu>