Fix filtering on IP heuristic subdissectors.
The IP dissector passes 'tree' to its heuristic dissectors instead
of the 'parent_tree' passed to the normally registered dissectors.
These two are generally the same except that 'tree' is set to NULL
in cases where IP is not referenced - a local optimization that
should not be passed down to the lower layers. The result is that
the filter:
ip && PROTO matches
but:
PROTO does not match.

This changes the two types of dissectors to behave the same by
getting 'parent_tree'.

Change-Id: I78690d4767d2e3a763fd4388792c49aa5df0a962
Reviewed-on: https://code.wireshark.org/review/8190
Reviewed-by: Guy Harris <guy@alum.mit.edu>