wsdg: clarify gerrit account requirement
A IRC user was asking about the meaning of `my.username` and noted that
information is scattered over places. This patch tries to make the
documentation more friendly for newcomers which are unfamiliar with
Gerrit (and maybe OpenID?).

The login page already recommends Launchpad, so no need to write it
here.

Change-Id: Ic55c9a2073d0045c6367f3efc8c842e53215f2dd
Reviewed-on: https://code.wireshark.org/review/8288
Reviewed-by: Mark <launchpad@markcunningham.ie>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Reviewed-by: Gerald Combs <gerald@wireshark.org>