Pull the MIMO control code into the beamforming report code.
Hopefully, that'll convince Microsoft's static analyzer that nr will
always be >= 1, so you can safely subtract 1 from it and use it as an
array index.

Get rid of the vht_mimo_control_t structure in favor of a few variables
for the bitfields we actually use.

Make some tables static - no need to initialize them every time we enter
the routine.

Change-Id: Icde05a768ea1a9c897b69003afcab1dddeffaaf5
Reviewed-on: https://code.wireshark.org/review/8411
Reviewed-by: Guy Harris <guy@alum.mit.edu>