Reorganize the NTLMSSP blob and AUTHENTICATE message parsing.
The "result" argument to dissect_ntlmssp_blob() is never null, so don't
check for it being null.

Have separate clauses for LmChallengeResponse and NtChallengeResponse,
and do the checks for NTLMv1 vs. NTLMv2 inside those clauses.

Do the copy to client_challenge within the AUTHENTICATE message parsing
only if we've already determined that it's an NTLMv2 message.

Add some comments to better explain what's being done and to ask some
questions.

Change-Id: I52345eaeac4252d928b2e477751817084bf4e363
Reviewed-on: https://code.wireshark.org/review/8523
Reviewed-by: Guy Harris <guy@alum.mit.edu>