Pop up error dialogs and quit if we don't have a working X11.
If XQuartz is missing the /usr/X11 symlink but the user doesn't allow us
to fix it, don't try to run Wireshark.

If the OS has X11 bundled but the user didn't install it, let them know,
and don't try to run Wireshark.

We should probably just say "your X11/XQuartz is hosed" for other
problem cases.

Get rid of some debugging fprintfs while we're at it.

Change-Id: I6ef4581e9a45967b4cda709c4be25dad41a9613d
Reviewed-on: https://code.wireshark.org/review/8598
Reviewed-by: Guy Harris <guy@alum.mit.edu>