Don't use the ServiceManagement framework when building for 10.5.
The 10.5 SDK doesn't have that framework, so we can't use it.

All three people for whom it's important that they be able to upgrade
their Leopard box, running Wireshark, release by release, all the way up
to Yosemite, and have Wireshark repair the damage done by the Yosemite
installer to the XQuartz installation a previous upgrade required, cried
bitter tears.

Change-Id: I9359da5ba8d5896d98fd3f15a14107822a65df70
Reviewed-on: https://code.wireshark.org/review/8601
Reviewed-by: Guy Harris <guy@alum.mit.edu>