Convert AFS dissector to use ptvcursor API.
The dissector was written in all macros, presumably to get just a single lines to display a field.  The ptvcursor API is good for that, and using it over macros more than halves the object size.  Real code (vs macros) is also much easier to use in a debugger.  It also makes it easier for the check* scripts to find possible errors.

Also eliminate proto_tree_add_text.

Change-Id: Id07e015b5a2d1a98a4b36e40a426442d826d9a09
Reviewed-on: https://code.wireshark.org/review/8723
Reviewed-by: Michael Mann <mmann78@netscape.net>
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>