WCCP: Fix up storing and decoding of IP addresses.
Try to decode as many IP addresses as we can, even if they don't fit in the
table we store them in.

Only add IP addresses in the table once.  We could theoretically reallocate
the table but the original code seems to assume the addresses should (in
non-fuzzed captures) only show up once per PDU.  This part of the change fixes
the fuzz failure.

Bug: 11153
Change-Id: I56b9854ac1342080c9f32699a3f97750fa335696
Reviewed-on: https://code.wireshark.org/review/8748
Petri-Dish: Jeff Morriss <jeff.morriss.ws@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>