Clean up --with-qt handling.
If you don't specify --with-qt or --without-qt, we build with Qt iff we
find it; we don't fail if we don't find it.

If you specify --with-qt, we build with Qt if we find it and fail if we
don't find it.

If you specify --without-qt, we don't look for Qt and don't build with
it.

This is all independent of --with-gtk2 or --with-gtk3.

Change-Id: I508d3281192bda9168fc46aba6011687c83ef818
Reviewed-on: https://code.wireshark.org/review/8861
Reviewed-by: Guy Harris <guy@alum.mit.edu>