Get rid of unnecessary stuff copied-and-pasted from the radiotap dissector.
What the dissector calls "radiotap" isn't even remotely like radiotap.
It took some flags from radiotap, but that's it.

Make the names of the flags in question match what's in the libwiretap
code.  Get rid of everything not in the libwiretap code, and get rid of
all code that does anything if those flags are set or that puts them
into the protocol tree.

This means that we don't do anything with the FCS or the Atheros
padding; just call the "802.11 with no FCS" dissector.

Change-Id: I4d3646d5d6c409b772985d0b05a86d5ee903157a
Reviewed-on: https://code.wireshark.org/review/8973
Reviewed-by: Guy Harris <guy@alum.mit.edu>