xip: add support for XIP packets
The eXpressive Internet Protocol (XIP) is the network layer
protocol for the eXpressive Internet Architecture (XIA), a
future Internet architecture project. The addresses in XIP are
directed acyclic graphs, so much of the code included in this
addition verifies the correctness of the DAGs and displays them
in human-readable form.

Bug: 11265
Change-Id: I948aaa73b927f8afc162d89689d184c5657f60b1
Reviewed-on: https://code.wireshark.org/review/8881
Reviewed-by: Evan Huus <eapache@gmail.com>
Petri-Dish: Evan Huus <eapache@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>