tools/pre-commit: compare files instead of mtime
Instead of unconditionally checking for the file mtime, compare file
contents instead. Perform this check only for invocations via a hook.

While at it, simplify setting defaults for the git directory and the
commit identifier. Support executions via relative paths and update the
help messages on top of the file.

Change-Id: I89c015fe6553ad52e07795683dd88cb219a26ed2
Reviewed-on: https://code.wireshark.org/review/8923
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>