Fix Windows CMake PortableApps dependencies
The nsis_uninstall target appears to create a temporary file
uninstaller_installer.exe that is renamed to uninstall.exe

The above rename upsets the portableapps_app_dir target that is
copying files from the CMake run dir into the portable apps dir.

The change adds a dependency on nsis_uninstall to the portableapps_package
target in an attempt to sequence the copy operation.

Change-Id: I9feb55599eff000f983c23ce29c89fbd3be0a6ac
Reviewed-on: https://code.wireshark.org/review/9153
Reviewed-by: Graham Bloice <graham.bloice@trihedral.com>
Petri-Dish: Graham Bloice <graham.bloice@trihedral.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>