If we don't know whether STBC is being used, we don't know NSTS.
If we don't know whether space-time block coding is being used, we don't
know the number of space-time streams, so don't try to calculate it and
don't put it into the protocol tree.

Mark it as generated, while we're at it, as it's not a value from the
radiotap header, and make it a guint.

Change-Id: Ib0b14f2f4f94e042d034311e10bb96b9b29a9d3e
Reviewed-on: https://code.wireshark.org/review/9206
Reviewed-by: Guy Harris <guy@alum.mit.edu>