If we don't know whether STBC is being used, we don't know NSTS.
If we don't know whether space-time block coding is being used, we don't
know the number of space-time streams, so don't try to calculate it and
don't put it into the protocol tree.

Mark it as generated, while we're at it, as it's not a value from the
radiotap header, and make it a guint.

Change-Id: If6e5db85948a61e92285a29d34a4da117dec55c7
Reviewed-on: https://code.wireshark.org/review/9207
Reviewed-by: Guy Harris <guy@alum.mit.edu>